+++ lib/src/clients/request.dart
@@ -8,7 +8,7 @@ import 'wrapper.dart';
 
 /// A client that overrides [url], [headers] of the request.
 ///
-/// Use [onJoinPath] to decide path joining strategy.
+/// Use [pathJoinStrategy] to decide path joining strategy.
 class RequestClient extends WrapperClient {
   /// If not null, this will merge with the url in the request.
   ///
@@ -19,8 +19,8 @@ class RequestClient extends WrapperClient {
   final Map<String, String>? headers;
 
   /// Decides path joining strategy.
-  /// Defaults to [PathJoinStrategy.originalOnlyIfHasHost].
-  final PathJoinCallback onJoinPath;
+  /// Defaults to [DefaultPathJoinStrategy].
+  final PathJoinStrategyCallback? pathJoinStrategy;
 
   /// Creates a [RequestClient] http client that can update the url
   /// and headers of a [BaseRequest] with [url], and [headers].
@@ -28,7 +28,7 @@ class RequestClient extends WrapperClient {
     Client client, {
     this.url,
     this.headers,
-    this.onJoinPath = PathJoinStrategy.originalOnlyIfHasHost,
+    this.pathJoinStrategy,
   }) : super(client);
 
   /// Returns an updated copy of [original] with the given [Request.body]
@@ -37,7 +37,13 @@ class RequestClient extends WrapperClient {
   @protected
   BaseRequest updateRequest(BaseRequest original) {
     if (url == null && (headers == null || headers!.isEmpty)) return original;
-    final newUrl = joinUrls(original.url, url, onJoinPath);
+    final newUrl = joinUrls(
+      original.url,
+      url,
+      PathJoinStrategy.onJoinPath(
+        pathJoinStrategy ?? (a, b) => PathJoinStrategy(a, b),
+      ),
+    );
     final hasEqualUrls = original.url.toString() == newUrl.toString();
     final hasEqualHeaders =
         const MapEquality().equals(headers, original.headers);